@@ -33,7 +33,7 @@ class ProfilesController < ApplicationController
 
 	def update
 		if @profile.update(profile_params)
-			redirect_to @profile, :notice => "Your Official Link is Successfully changed. Share it with your Friends and Everyone. You can change it Again."
+			redirect_to @profile, :notice => "Your Profile Information is changed. You can change it again if you want."
 		else
 			render 'edit'
 		end
@@ -46,6 +46,6 @@ class ProfilesController < ApplicationController
 	end
 
 	def profile_params
-		params.require(:profile).permit(:line1, :line2, :slug)
+		params.require(:profile).permit(:line1, :line2, :line3, :line4, :line5, :line6, :line7, :line8, :line9, :line10, :line11, :line12, :line13, :line14, :line15, :line16, :slug)
 	end
 end
