@@ -3,8 +3,7 @@ class Profile < ApplicationRecord
 	friendly_id :slug_candidates, use: :slugged
   def slug_candidates
     [
-      :line1,
-      [:line1, :id]
+      :id
     ]
   end
 	def should_generate_new_friendly_id?
@@ -12,5 +11,7 @@ class Profile < ApplicationRecord
   end
   belongs_to :user
   has_many :posts, -> { order "created_at DESC" }, dependent: :destroy
-  validates :line1, presence: true, uniqueness: true, length: { maximum: 48 }, format: { with: /\A[a-z0-9_]+\Z/ }
+  validates :line1, uniqueness: true, length: { maximum: 48 }, format: { with: /\A[a-z0-9_]+\Z/ }
+  include PgSearch
+  pg_search_scope :search, :against => [:line2, :line3, :line11, :line12, :line13]
 end
