@@ -2,15 +2,15 @@
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
 #
-# Note that this schema.rb definition is the authoritative source for your
-# database schema. If you need to create the application database on another
-# system, you should be using db:schema:load, not running all the migrations
-# from scratch. The latter is a flawed and unsustainable approach (the more migrations
-# you'll amass, the slower it'll run and the greater likelihood for issues).
+# This file is the source Rails uses to define your schema when running `bin/rails
+# db:schema:load`. When creating a new database, `bin/rails db:schema:load` tends to
+# be faster and is potentially less error prone than running all of your
+# migrations from scratch. Old migrations may fail to apply correctly if those
+# migrations use external dependencies or application code.
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20171108183444) do
+ActiveRecord::Schema.define(version: 2021_03_13_092006) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -59,7 +59,22 @@ ActiveRecord::Schema.define(version: 20171108183444) do
     t.integer "user_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.string "image"
     t.string "slug"
+    t.string "line3"
+    t.string "line4"
+    t.string "line5"
+    t.string "line6"
+    t.string "line7"
+    t.string "line8"
+    t.string "line9"
+    t.string "line10"
+    t.text "line11"
+    t.text "line12"
+    t.text "line13"
+    t.text "line14"
+    t.text "line15"
+    t.text "line16"
     t.index ["slug"], name: "index_profiles_on_slug"
   end
 
